+++ src/environments/environment.prod.ts
@@ -1,7 +1,7 @@
 export const environment = {
   production: true,
-  appVersion: '0.2.1',
-  releaseDate: '2021-11-17 23:30',
+  appVersion: '0.2.2',
+  releaseDate: '2021-11-18 23:30',
   baseUrl: 'https://info-rincon.web.app/',
   setAudit: true,
   firebase: {